Added some more market-dex routes, implemented migration system

+++ app/database.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+import sys, os
 from peewee import *
 from playhouse.postgres_ext import *
 import redis
@@ -17,7 +19,30 @@ TABLES = [
     MarketItemPricePoint
 ]
 
+def migrate(module):
+    print "Running migration for %s" % module.__name__
+    
+    module.pre()
+    module.run()
+    module.post()
+
 if __name__ == "__main__":
-    for table in TABLES:
-        table.drop_table(True, cascade=True)
-        table.create_table(True)
+    if len(sys.argv) <= 1:
+        print "Usage: ./database.py [build/reset/migrate]"
+        sys.exit(1)
+
+    if sys.argv[1] == "build":
+        print "Building new tables..."
+        map(lambda i: i.create_table(True), TABLES)
+
+    if sys.argv[1] == "reset":
+        print "Resetting DB..."
+        for table in TABLES:
+            table.drop_table(True, cascade=True)
+            table.create_table(True)
+
+    if sys.argv[1] == "migrate":
+        m = "migrations.%s" % sys.argv[2]
+        __import__(m)
+        migrate(sys.modules[m])
+
